Move utf8(De|En)code to new text module
