+++ src/main/scala/scalaz/stream/process1.scala
@@ -1,7 +1,6 @@
 package scalaz.stream
 
 import collection.immutable.Vector
-import java.nio.charset.Charset
 
 import scalaz.{\/, -\/, \/-, Monoid, Semigroup, Equal, Order}
 import scalaz.\/._
@@ -13,7 +12,6 @@ import scalaz.stream.Process.Emit
 import scala.Some
 import scalaz.stream.Process.Halt
 import scalaz.stream.ReceiveY.{HaltL, HaltR, ReceiveR, ReceiveL}
-import scodec.bits.ByteVector
 
 trait process1 {
 
@@ -549,57 +547,6 @@ trait process1 {
   /** Wraps all inputs in `Some`, then outputs a single `None` before halting. */
   def terminated[A]: Process1[A,Option[A]] =
     lift[A,Option[A]](Some(_)) ++ emit(None)
-
-  private val utf8Charset = Charset.forName("UTF-8")
-
-  /** Converts UTF-8 encoded `ByteVector` into `String`. */
-  val utf8Decode: Process1[ByteVector,String] = {
-    /**
-     * Returns the number of continuation bytes if `b` is an ASCII byte or a
-     * leading byte of a multi-byte sequence, and -1 otherwise.
-     */
-    def continuationBytes(b: Byte): Int = {
-      if      ((b & 0x80) == 0x00) 0 // ASCII byte
-      else if ((b & 0xE0) == 0xC0) 1 // leading byte of a 2 byte seq
-      else if ((b & 0xF0) == 0xE0) 2 // leading byte of a 3 byte seq
-      else if ((b & 0xF8) == 0xF0) 3 // leading byte of a 4 byte seq
-      else -1                        // continuation byte or garbage
-    }
-
-    /**
-     * Returns the length of an incomplete multi-byte sequence at the end of
-     * `bs`. If `bs` ends with an ASCII byte or a complete multi-byte sequence,
-     * 0 is returned.
-     */
-    def lastIncompleteBytes(bs: ByteVector): Int = {
-      val lastThree = bs.toIndexedSeq.reverseIterator.take(3)
-      lastThree.map(continuationBytes).zipWithIndex.find {
-        case (c, _) => c >= 0
-      } map {
-        case (c, i) => if (c == i) 0 else i + 1
-      } getOrElse(0)
-    }
-
-    def splitAtLastIncompleteChar(bs: ByteVector): (Option[ByteVector], Option[ByteVector]) = {
-      val splitIndex = bs.length - lastIncompleteBytes(bs)
-
-      if (bs.isEmpty || splitIndex == bs.length)
-        (Some(bs), None)
-      else if (splitIndex == 0)
-        (None, Some(bs))
-      else {
-        val (complete, rest) = bs.splitAt(splitIndex)
-        (Some(complete), Some(rest))
-      }
-    }
-
-    repartition2(splitAtLastIncompleteChar)
-      .map(bs => new String(bs.toArray, utf8Charset))
-  }
-
-  /** Convert `String` inputs to UTF-8 encoded `ByteVector`. */
-  val utf8Encode: Process1[String,ByteVector] =
-    lift(s => ByteVector.view(s.getBytes(utf8Charset)))
 
   /**
    * Outputs a sliding window of size `n` onto the input.
