+++ src/test/scala/scalaz/stream/Utf8DecodeSpec.scala
@@ -8,9 +8,9 @@ import scalaz.syntax.equal._
 import scodec.bits.ByteVector
 
 import Process._
-import process1._
+import text._
 
-object Utf8DecodeSpec extends Properties("process1.utf8Decode") {
+object Utf8DecodeSpec extends Properties("text.utf8Decode") {
 
   def utf8Bytes(a: Array[Int]): ByteVector = ByteVector(a.map(_.toByte))
   def utf8Bytes(c: Char): ByteVector = utf8Bytes(c.toString)
